# docs/openapi.yaml
openapi: 3.0.3
info:
  title: KOR Farm v2 API
  version: 0.1.0
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
...
  /v1/files/presign:
    post:
      tags: [Files]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PresignRequest" }
      responses:
        "200": { $ref: "#/components/responses/PresignResponse" }
  /v1/files/{fileId}/download:
    get:
      tags: [Files]
      parameters:
        - in: path
          name: fileId
          required: true
          schema: { type: string }
      responses:
        "200": { $ref: "#/components/responses/FileDownloadResponse" }
...
  /v1/shop/products/{productId}:
    get:
      tags: [Shop]
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: string }
      responses:
        "200": { $ref: "#/components/responses/ProductDetailResponse" }
  /v1/shop/orders:
    post:
      tags: [Shop]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrderCreateRequest" }
      responses:
        "200": { $ref: "#/components/responses/OrderCreateResponse" }
    get:
      tags: [Shop]
      responses:
        "200": { $ref: "#/components/responses/OrderListResponse" }
...
  /v1/admin/shop/products:
    get:
      tags: [Admin]
      responses:
        "200": { $ref: "#/components/responses/AdminProductListResponse" }
    post:
      tags: [Admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminProductRequest" }
      responses:
        "200": { $ref: "#/components/responses/AdminProductResponse" }
  /v1/admin/shop/products/{productId}:
    patch:
      tags: [Admin]
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminProductRequest" }
      responses:
        "200": { $ref: "#/components/responses/AdminProductResponse" }
    delete:
      tags: [Admin]
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: string }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Ok:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiResponse"
...
    AdminProductResponse:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AdminProductResponse"
    AdminProductListResponse:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AdminProductListResponse"
...
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success: { type: boolean }
        data:
          oneOf:
            - type: object
              additionalProperties: true
            - type: array
              items: {}
        error: { $ref: "#/components/schemas/Error" }
        request_id: { type: string }
      required: [success]
    Error:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
...
    PresignRequest:
      type: object
      properties:
        purpose: { type: string }
        filename: { type: string }
        mime: { type: string }
        size: { type: integer }
      required: [purpose, filename, mime, size]
...
    Product:
      type: object
      properties:
        product_id: { type: string }
        name: { type: string }
        price: { type: integer }
        stock: { type: integer }
        status: { type: string }
...
    OrderItemRequest:
      type: object
      properties:
        product_id: { type: string }
        quantity: { type: integer }
      required: [product_id, quantity]
    OrderCreateRequest:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItemRequest" }
        address: { type: object }
      required: [items, address]
...
    AdminProductListResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              type: array
              items: { $ref: "#/components/schemas/Product" }
...
    AdminProductRequest:
      type: object
      properties:
        name: { type: string }
        price: { type: integer }
        stock: { type: integer }
        status: { type: string }
      required: [name, price]
    AdminProductResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/Product" }
...